many tweaks
Switched the info available on F2 and F3 so info fetch for the corresponding puzzle happens on the same floor. Created more dorm objects to interact with and reworked ending sequence to display Carson's replies on a phone interface. Ruby's reactions to 4th floor emails. Fixed background colors for text in flag book. Transitioned MacLeod history book to a plaque.