"Remembering" version
Moved Courtyard to trigger off a missing library history exhibit. Rewrote Courtyard and Lost Woods to characterize them as visits to Ruby's "mind dungeon" to recall information that's unavailable or erased in the library's physical environment, and characterize each puzzle as an internal quest for a key bit of information.